requirements.txt
azure-ai-voicelive[aiohttp]>=1.1.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
